+++ docs/examples/modal/ExProgrammatic.vue
@@ -2,16 +2,18 @@
     <section>
         <button class="button is-primary is-medium"
             @click="imageModal()">
-            Launch image modal
+            Launch image modal (HTML)
         </button>
         <button class="button is-primary is-medium"
             @click="cardModal()">
-            Launch card modal
+            Launch card modal (Component)
         </button>
     </section>
 </template>
 
 <script>
+    import ModalForm from './ModalForm'
+
     export default {
         methods: {
             imageModal() {
@@ -23,35 +25,9 @@
             },
             cardModal() {
                 this.$modal.open({
-                    width: 640,
-                    content: `<div class="card">
-                        <div class="card-image">
-                            <figure class="image is-4by3">
-                            <img src="./static/img/placeholder-1280x960.png" alt="Image">
-                            </figure>
-                        </div>
-                        <div class="card-content">
-                            <div class="media">
-                            <div class="media-left">
-                                <figure class="image is-48x48">
-                                <img src="./static/img/placeholder-1280x960.png" alt="Image">
-                                </figure>
-                            </div>
-                            <div class="media-content">
-                                <p class="title is-4">John Smith</p>
-                                <p class="subtitle is-6">@johnsmith</p>
-                            </div>
-                            </div>
-
-                            <div class="content">
-                            Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                            Phasellus nec iaculis mauris. <a>@bulmaio</a>.
-                            <a>#css</a> <a>#responsive</a>
-                            <br>
-                            <small>11:09 PM - 1 Jan 2016</small>
-                            </div>
-                        </div>
-                    </div>`
+                    parent: this,
+                    component: ModalForm,
+                    hasModalCard: true
                 })
             }
         }